+++ app/Http/Controllers/Admin/VendorRequestController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\DataTables\VendorRequestsDataTable;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -96,6 +97,9 @@ class VendorRequestController extends Controller
     public function destroy(string $id)
     {
         //dd($id);
+        if (Auth::user()->id != 1) {
+            abort(404);
+        }
         $request = User::findOrFail($id);
         if(File::exists(public_path($request->document))){
             File::delete(public_path($request->document));
